# docs/swagger.json
{
   "swagger":"2.0",
   "info":{
      "description":"Manage Craft Beers using Spring Boot REST API",
      "version":"1.0.0",
      "title":"Craft Beers Spring Boot REST API",
      "contact":{
         "name":"Edivaldo Jr.",
         "url":"https://juniormascarenhas.com",
         "email":"contato@juniormascarenhas.com"
      },
      "license":{
         "name":"Apache License Version 2.0",
         "url":"https://www.apache.org/licenses/LICENSE-2.0"
      }
   },
   "host":"localhost:8080",
   "basePath":"/",
   "tags":[
      {
         "name":"craft-beer-controller",
         "description":"Craft Beer Controller"
      }
   ],
   "paths":{
      "/craft-beers":{
         "get":{
            "tags":[
               "craft-beer-controller"
            ],
            "summary":"getAll",
            "operationId":"getAllUsingGET",
            "produces":[
               "application/json"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/CraftBeer"
                     }
                  }
               },
               "204":{
                  "description":"No Content"
               },
               "404":{
                  "description":"Not Found!"
               },
               "500":{
                  "description":"500 message"
               }
            },
            "deprecated":false
         },
         "post":{
            "tags":[
               "craft-beer-controller"
            ],
            "summary":"save",
            "operationId":"saveUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"craftBeer",
                  "description":"craftBeer",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/CraftBeer"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/CraftBeer"
                  }
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "craft-beer-controller"
            ],
            "summary":"deleteAllCraftBeer",
            "operationId":"deleteAllCraftBeerUsingDELETE",
            "produces":[
               "application/json"
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"object"
                  }
               }
            },
            "deprecated":false
         }
      },
      "/craft-beers/name/{name}":{
         "get":{
            "tags":[
               "craft-beer-controller"
            ],
            "summary":"findByName",
            "operationId":"findByNameUsingGET",
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "name":"name",
                  "in":"path",
                  "description":"name",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"array",
                     "items":{
                        "$ref":"#/definitions/CraftBeer"
                     }
                  }
               },
               "204":{
                  "description":"No Content"
               },
               "404":{
                  "description":"Not Found!"
               },
               "500":{
                  "description":"500 message"
               }
            },
            "deprecated":false
         }
      },
      "/craft-beers/{id}":{
         "get":{
            "tags":[
               "craft-beer-controller"
            ],
            "summary":"getCraftBeerById",
            "operationId":"getCraftBeerByIdUsingGET",
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/CraftBeer"
                  }
               },
               "204":{
                  "description":"No Content"
               },
               "404":{
                  "description":"Not Found!"
               },
               "500":{
                  "description":"500 message"
               }
            },
            "deprecated":false
         },
         "put":{
            "tags":[
               "craft-beer-controller"
            ],
            "summary":"updateCraftBeer",
            "operationId":"updateCraftBeerUsingPUT",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"craftBeer",
                  "description":"craftBeer",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/CraftBeer"
                  }
               },
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/CraftBeer"
                  }
               }
            },
            "deprecated":false
         },
         "delete":{
            "tags":[
               "craft-beer-controller"
            ],
            "summary":"deleteCraftBeer",
            "operationId":"deleteCraftBeerUsingDELETE",
            "produces":[
               "application/json"
            ],
            "parameters":[
               {
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int64"
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "type":"object"
                  }
               }
            },
            "deprecated":false
         }
      }
   },
   "definitions":{
      "CraftBeer":{
         "type":"object",
         "properties":{
            "abv":{
               "type":"number",
               "format":"double"
            },
            "breweryName":{
               "type":"string"
            },
            "categoryName":{
               "type":"string"
            },
            "ibu":{
               "type":"integer",
               "format":"int32"
            },
            "id":{
               "type":"integer",
               "format":"int64"
            },
            "name":{
               "type":"string"
            }
         },
         "title":"CraftBeer"
      }
   }
}